# PERF: improve performance of Point(x, y) constructor