This avoids creating a separate array for the x and y argument, and then stacking both arrays (this is what happens in the `points(..)` under the hood, as that can both accept a single 2D array of separate x/y arrays):

```
In [2]: %timeit Point(1, 1)
12.6 µs ± 442 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)    # <-- main
4.34 µs ± 60.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)   # <-- PR
```